backend/requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
httpx==0.28.1
python-dotenv==1.2.2
transformers>=4.34.0
torch
accelerate
safetensors
huggingface_hub
motor
pymongo
